+++ rbc/_omnisci/numpy_funcs.py
