+++ rbc/_omnisci/omnisci_array.py
