+++ rbc/_omnisci/python_operators.py
